+++ apps/oauth-mock-server/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @stackframe/oauth-mock-server
 
+## 2.6.1
+
+### Patch Changes
+
+- @stackframe/stack-shared@2.6.1
+
 ## 2.6.0
 
 ### Minor Changes