+++ examples/docs-examples/CHANGELOG.md
@@ -1,5 +1,14 @@
 # demo-app
 
+## 2.6.1
+
+### Patch Changes
+
+- Updated dependencies
+  - @stackframe/stack@2.6.1
+  - @stackframe/stack-shared@2.6.1
+  - @stackframe/stack-ui@2.6.1
+
 ## 2.6.0
 
 ### Minor Changes